fix: optimization code
