+++ src/screens/PassingQuiz/components/QuizList/styles.ts
