+++ src/screens/PassingQuiz/components/QuizProcess/styles.ts
