+++ src/screens/PassingQuiz/components/QuizResultScreen/styles.ts
